text-generation/requirements.txt
google-generativeai
fastapi
uvicorn